Core: Refactoring of the actual API call
